requirements.txt
gradio
torch
torchvision
numpy
Pillow
bidict
huggingface-hub
git-lfs